CRJ-buxfixes Adds fixes in core services - Refactor in core/services fixing a bug where some error was thrown out
