+++ src/app/core/services/models/food.class.ts
@@ -1,0 +1,11 @@
+import { FoodTime, FoodType } from './food.enum';
+
+export class FoodTable {
+  foodTime: string;
+  foodType: string;
+
+  constructor(fdRw: any) {
+    this.foodTime = FoodTime[fdRw.foodTime];
+    this.foodType = FoodType[fdRw.foodType];
+  }
+}
