+++ src/app/core/services/models/food.enum.ts
@@ -13,13 +13,3 @@ export enum FoodType {
   fruits = 'Frutas',
   dairy = 'Leche'
 }
-
-export class FoodTable {
-  foodTime: string;
-  foodType: string;
-
-  constructor(fdRw: any) {
-    this.foodTime = FoodTime[fdRw.foodTime];
-    this.foodType = FoodType[fdRw.foodType];
-  }
-}
